+++ .github/CHANGELOG.md
@@ -12,3 +12,6 @@
 11. [Game] Enhancement - Added basketball court building
 12. [Game] Test - Added tests to check Event Handler class
 13. [Game] Enhancement - Added leaderboard upon game over (in progress)
+13. [Game] Test - Added tests to check GameUtils class
+14. [Game] Test - Added tests to check Timer class 
+15. [Game] Test - Added tests to check Main class